Move action creators and reducers into a directory structure that mimics the state tree.
