--- client/state/reducers/index.js
@@ -1,10 +1,0 @@
-import { combineReducers } from 'redux';
-import settings from './settings';
-import form from './form';
-
-const rootReducer = combineReducers( {
-	settings,
-	form,
-} );
-
-export default rootReducer;
