--- client/state/store/index.js
@@ -1,21 +1,0 @@
-import { createStore } from 'redux';
-import rootReducer from '../reducers';
-
-const configureStore = ( initialState ) => {
-	const store = createStore(
-		rootReducer,
-		initialState,
-		window.devToolsExtension ? window.devToolsExtension() : f => f
-	);
-
-	if ( module.hot ) {
-		module.hot.accept( '../reducers', () => {
-			const nextRootReducer = require( '../reducers' ).default;
-			store.replaceReducer( nextRootReducer );
-		} );
-	}
-
-	return store;
-};
-
-export default configureStore;
